# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASAPhoto REST APIs
  title: WASAPhoto APIs
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
security:
  - BasicAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    banninguserid:
      schema:
        description: unique id related to user who bans
        type: integer
        example: 1
        readOnly: true
      name: banninguserid
      in: path
      required: true
      description: User ID who bans
    banneduserid:
      schema:
        description: unique id related to user who banned
        type: integer
        example: 1
        readOnly: true
      name: banneduserid
      in: path
      required: true
      description: User ID who bans
...
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    ForbiddenRequest:
      description: |-
        Clien executes not request to the contents without
        required permissions.    
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedRequest:
      description: |-
        The request had been executed without login
...
  schemas:
...
    banToUser:
      title: ban user
      description: ban user entity
      type: object
      properties:
        banningUser: 
          description: Id of user who bans
          type: integer
          example: 2
        bannedUser:
          description: Id of banned user
          type: integer
          example: 7
      example:
        {
          banningUser: 2,
          bannedUser: 1
        }
        
    #there must be a request bodies for all components
...
    Profile:
      title: Profile
      description: User represented as an object
      type: object
      properties:
        avatar:
          description: User's profile image
          type: string
          format: binary
        userid:
          description: Unique user identifier
          type: integer
          example: 1
          readOnly: true
        username:
          description: Unique username identifier
          type: string
          example: username1
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
        password:
          description: password to enter inside application
          type: string
          example: "*zafgz3lU341"
          pattern: '^.*?$'
          minLength: 8
          maxLength: 16
        quantitySubscribers:
          description: number of subscribers to the current account
          type: string
          example: "3k"
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
        quantitySubscriptions:
          description: number of users current user subscribed to
          type: string
          example: "2k"
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
          
      example:
        {
          avatar: "binary string",
          userid: 1,
          username: 'User1',
          quantitySubscribers: '2.3k',
          quantitySubscriptions: '34m'
        }
      
          
#there are some paths inside which u can check below          
paths:
...
  /profiles/{banninguserid}/banuser/{banneduserid}:
  
    parameters:
      - $ref: "#/components/parameters/banninguserid"
      - $ref: "#/components/parameters/banneduserid"
    put:
      operationId: banUser
      tags:
      - Account
      summary: ban User without giving possibility to see posts
      description: |-
        addition banned user to collection of banned users by a certain one
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/banToUser" }
      responses:
        "204":
           description: Ban entity had been added successfully
          
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: unbanUser
      tags:
        - Account
      security:
        - BearerAuth: []
      summary: unban User and give possibility to see posts
      description: |-
        Removes a pair (userBans, userBanned) from a database
        after that operation banned user can see posts of userBans
      responses:
        "204":
          description: Ban entity was deleted successfully
#        "200":
#          description: Entity for banning had been deleted
#          content:
#            application/json:
#              schema: { $ref: "#/components/schemas/userBanChanged" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{banninguserid}/banuser:
    parameters:
      - $ref: "#/components/parameters/banninguserid"

    get:
      operationId: bannedUsers
      tags:
      - Account
      summary: shows list of banned users
      
         
      description: |-
        Shows list of banned users by current one
      responses:
        "200":
          description: list of banned users returned
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 10
                description: Profiles blocked some time ago
                items: { $ref: "#/components/schemas/Profile" }
              example: 
                [{
                    avatar: "binary string",
                    userid: 1,
                    username: 'User1',
                    quantitySubscribers: '2.9k',
                    quantitySubscriptions: '64m'
                  },
                  {
                      avatar: "binary string",
                      userid: 2,
                      username: 'User2',
                      quantitySubscribers: '7.3k',
                      quantitySubscriptions: '14m'
                    }






                  ]




              
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    
        
  # i am not sure if the method below should exist

  #/profiles/me/settings:
  #  get:
  #    operationId: listProfileSettings
  #    summary: Returns settings for profile
  #    description: |-
  #      Retrieve all settings from profile as list
  #    responses:
  #      "200":
  #        description: returns all settings related to current user
  #        content:
  #          application/json:
  #            schema: { $ref: "#/components/schemas/Comment" }
  #      "400": { $ref: "#/components/responses/BadRequest" }
  #      "403": { $ref: "#/components/responses/ForbiddenRequest" }
  #      "500": { $ref: "#/components/responses/InternalServerError" }
...